requirements.txt
playwright==1.42.0
pandas==2.2.0
genderize==0.3.1
openpyxl==3.1.2
python-dotenv==1.0.1
pdfplumber==0.10.3
beautifulsoup4==4.12.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.119.0
requests==2.31.0
pyarrow==15.0.0
numpy==1.26.4
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-playwright==0.4.4
pytest-asyncio==0.23.5
responses==0.24.1